# assignment4/openapi.yaml
openapi: 3.0.3
info:
  title: CampusEats Orders API
  version: 1.0.0
  description: Assignment 4. In-memory teaching service. All failures use Problem;
    framework 405/413/500 also use this shape.
servers:
- url: http://127.0.0.1:5000
paths:
  /orders:
    post:
      operationId: createOrder
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          headers:
            Location:
              description: Created resource URI
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/Error400'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '503':
          $ref: '#/components/responses/Error503'
        default:
          description: Other HTTP errors (including 405, 413 and 500) use the same
            problem representation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          $ref: '#/components/schemas/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
    get:
      operationId: listOrders
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '400':
          $ref: '#/components/responses/Error400'
        default:
          description: Other HTTP errors (including 405, 413 and 500) use the same
            problem representation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - in: query
        name: status
        schema:
          $ref: '#/components/schemas/Status'
      - in: query
        name: user_id
        schema:
          $ref: '#/components/schemas/Identifier'
  /orders/{order_id}:
    get:
      operationId: getOrder
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        default:
          description: Other HTTP errors (including 405, 413 and 500) use the same
            problem representation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - &id001
        in: path
        name: order_id
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
  /orders/{order_id}/cancellation:
    post:
      operationId: cancelOrder
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cancellation'
          headers:
            Location:
              description: Created resource URI
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        default:
          description: Other HTTP errors (including 405, 413 and 500) use the same
            problem representation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancellationRequest'
    get:
      operationId: getCancellation
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cancellation'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        default:
          description: Other HTTP errors (including 405, 413 and 500) use the same
            problem representation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - *id001
components:
  schemas:
    Identifier:
      type: string
      format: uuid
    IdempotencyKey:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^[A-Za-z0-9._:-]+$
    Status:
      type: string
      enum:
      - PLACED
      - ACCEPTED
      - PREPARING
      - READY
      - OUT_FOR_DELIVERY
      - DELIVERED
      - COMPLETED
      - CANCELLED
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    Item:
      type: object
      additionalProperties: false
      required:
      - item_id
      - item_name_snapshot
      - quantity
      - unit_price_snapshot
      properties:
        item_id:
          $ref: '#/components/schemas/Identifier'
        item_name_snapshot:
          type: string
          minLength: 1
          maxLength: 120
        quantity:
          type: integer
          minimum: 1
          maximum: 100
        unit_price_snapshot:
          type: string
          pattern: ^\d{1,6}\.\d{2}$
    CreateOrder:
      type: object
      additionalProperties: false
      required:
      - user_id
      - vendor_id
      - fulfillment_type
      - items
      - payment_token
      properties:
        user_id:
          $ref: '#/components/schemas/Identifier'
        vendor_id:
          $ref: '#/components/schemas/Identifier'
        fulfillment_type:
          type: string
          enum:
          - delivery
          - pickup
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/Item'
        payment_token:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true
    CancellationRequest:
      type: object
      additionalProperties: false
      required:
      - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 200
    Cancellation:
      type: object
      required:
      - order_id
      - reason
      - cancelled_at
      properties:
        order_id:
          $ref: '#/components/schemas/Identifier'
        reason:
          type: string
        cancelled_at:
          type: string
          format: date-time
    Order:
      type: object
      required:
      - order_id
      - user_id
      - vendor_id
      - status
      - total_amount
      - fulfillment_type
      - items
      - created_at
      properties:
        order_id:
          $ref: '#/components/schemas/Identifier'
        user_id:
          $ref: '#/components/schemas/Identifier'
        vendor_id:
          $ref: '#/components/schemas/Identifier'
        status:
          $ref: '#/components/schemas/Status'
        total_amount:
          type: string
          pattern: ^\d{1,6}\.\d{2}$
        fulfillment_type:
          type: string
          enum:
          - delivery
          - pickup
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        created_at:
          type: string
          format: date-time
    OrderList:
      type: object
      required:
      - orders
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    Location:
      type: string
      format: uri-reference
  responses:
    Error400:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error404:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error409:
      description: State or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error415:
      description: JSON required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error422:
      description: Domain refusal, including PAYMENT_DECLINED (Assignment 3 GatewayFault
        CARD_DECLINED)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error503:
      description: 'PAYMENT_UNAVAILABLE: dependency unavailable or invalid response'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'